requirement-scrape.txt
geopandas
streetview
python-dotenv

fastapi
uvicorn